\name{print.express.certificate}
\alias{print.express.certificate}
\title{Print method for express certificates}
\description{
Print method for express certificates objects
}
\usage{
\method{print}{express.certificate}(x, digits = max(3, getOption("digits") - 3), ...)
}

\arguments{
  \item{x}{An object of S3 class "express.certificate"}
  \item{digits}{Number of digits for printing the object "express.certificate" in method 
                \code{print.express.certificate}}
  \item{\dots}{further arguments passed to or from other methods}
}
\details{
The method \code{print.express.certificate} can be used for pretty printing of express certificates properties.
}

\author{Stefan Wilhelm \email{wilhelm@financial.com}}
